+++ src/lib.rs
@@ -12,24 +12,19 @@
 //! let events_loop = glutin::EventsLoop::new();
 //! let window = glutin::WindowBuilder::new()
 //!     .with_title("Hello world!")
-//!     .with_dimensions(1024, 768);
+//!     .with_dimensions(glutin::dpi::LogicalSize::new(1024.0, 768.0));
 //! let context = glutin::ContextBuilder::new();
 //! let gl_window = glutin::GlWindow::new(window, context, &events_loop).unwrap();
 //! # }
 //! ```
 //!
-//! For contexts that are *not* associated with any particular window, see the HeadlessContext
-//! type.
-//!
 //! # Features
 //!
-//! This crate has two Cargo features: `window` and `headless`.
+//! This crate has one Cargo feature: `window`.
 //!
 //!  - `window` allows you to create regular windows and enables the `WindowBuilder` object.
-//!  - `headless` allows you to do headless rendering, and enables
-//!     the `HeadlessRendererBuilder` object.
 //!
-//! By default only `window` is enabled.
+//! By default `window` is enabled.
 
 #[cfg(target_os = "windows")]
 #[macro_use]
@@ -61,16 +56,15 @@ extern crate x11_dl;
 #[cfg(any(target_os = "linux", target_os = "freebsd", target_os = "dragonfly", target_os = "openbsd"))]
 extern crate wayland_client;
 
-pub use headless::{HeadlessRendererBuilder, HeadlessContext};
 pub use winit::{
     AvailableMonitorsIter,
     AxisId,
     ButtonId,
     ControlFlow,
     CreationError as WindowCreationError,
-    CursorState,
     DeviceEvent,
     DeviceId,
+    dpi,
     ElementState,
     Event,
     EventsLoop,
@@ -98,12 +92,14 @@ use std::io;
 
 mod api;
 mod platform;
-mod headless;
 
 pub mod os;
 
 /// A trait for types associated with a GL context.
-pub trait GlContext {
+pub trait GlContext
+where
+    Self: Sized,
+{
     /// Sets the context as the current context.
     unsafe fn make_current(&self) -> Result<(), ContextError>;
 
@@ -113,30 +109,8 @@ pub trait GlContext {
     /// Returns the address of an OpenGL function.
     fn get_proc_address(&self, addr: &str) -> *const ();
 
-    /// Swaps the buffers in case of double or triple buffering.
-    ///
-    /// You should call this function every time you have finished rendering, or the image may not
-    /// be displayed on the screen.
-    ///
-    /// **Warning**: if you enabled vsync, this function will block until the next time the screen
-    /// is refreshed. However drivers can choose to override your vsync settings, which means that
-    /// you can't know in advance whether `swap_buffers` will block or not.
-    fn swap_buffers(&self) -> Result<(), ContextError>;
-
     /// Returns the OpenGL API being used.
     fn get_api(&self) -> Api;
-
-    /// Returns the pixel format of the main framebuffer of the context.
-    fn get_pixel_format(&self) -> PixelFormat;
-
-    /// Resize the GL context.
-    ///
-    /// Some platforms (macos, wayland) require being manually updated when their window or
-    /// surface is resized.
-    ///
-    /// The easiest way of doing this is to call this method for each `Resized` window event that
-    /// is received with the width and height given by the event.
-    fn resize(&self, width: u32, height: u32);
 }
 
 /// Represents an OpenGL context.
@@ -318,15 +292,111 @@ impl<'a> ContextBuilder<'a> {
         self.pf_reqs.srgb = srgb_enabled;
         self
     }
+
+    /// Sets whether double buffering should be enabled.
+    ///
+    /// The default value is `None`.
+    ///
+    /// ## Platform-specific
+    ///
+    /// This option will be taken into account on the following platforms:
+    ///
+    ///   * MacOS
+    ///   * Linux using GLX with X
+    ///   * Windows using WGL
+    ///
+    #[inline]
+    pub fn with_double_buffer(mut self, double_buffer: Option<bool>) -> Self {
+        self.pf_reqs.double_buffer = double_buffer;
+        self
+    }
+
+    /// Sets whether hardware acceleration is required.
+    ///
+    /// The default value is `Some(true)`
+    ///
+    /// ## Platform-specific
+    ///
+    /// This option will be taken into account on the following platforms:
+    ///
+    ///   * MacOS
+    ///   * Linux using EGL with either X or Wayland
+    ///   * Windows using EGL or WGL
+    ///   * Android using EGL
+    ///
+    #[inline]
+    pub fn with_hardware_acceleration(mut self, acceleration: Option<bool>) -> Self {
+        self.pf_reqs.hardware_accelerated = acceleration;
+        self
+    }
 }
 
 impl GlWindow {
     /// Builds the given window along with the associated GL context, returning the pair as a
     /// `GlWindow`.
     ///
-    /// Error should be very rare and only occur in case of permission denied, incompatible system,
-    /// out of memory, etc.
+    /// The context made can be shared with:
+    ///  - headless contexts made with the `shareable_with_windowed_contexts`
+    ///  flag set to `true`; and
+    ///  - contexts made when creating a `GlWindow`.
+    ///
+    /// You are not guaranteed to receive an error if you share a context with an
+    /// other context which you're not permitted to share it with, as according
+    /// to:
+    ///  - the restrictions stated by us above; and
+    ///  - the restrictions imposed on you by the platform your application runs
+    ///  on. (Please refer to `README-SHARING.md`)
+    ///
+    /// Failing to follow all the context sharing restrictions imposed on you
+    /// may result in unsafe behavior.
+    ///
+    /// This safe variant of `new_shared` will panic if you try to share it with
+    /// an existing context.
+    ///
+    /// Error should be very rare and only occur in case of permission denied,
+    /// incompatible system out of memory, etc.
     pub fn new(
+        window_builder: WindowBuilder,
+        context_builder: ContextBuilder,
+        events_loop: &EventsLoop,
+    ) -> Result<Self, CreationError>
+    {
+        let ContextBuilder { pf_reqs, gl_attr } = context_builder;
+        let gl_attr = gl_attr.map_sharing(|_ctxt| panic!("Context sharing is not allowed when using `new()`. Please instead use `new_shared()`."));
+        
+        // Not all platforms support context sharing yet, when they do, their 
+        // `new.*` functions should be marked unsafe.
+        #[allow(unused_unsafe)] 
+        unsafe {
+            platform::Context::new(window_builder, events_loop, &pf_reqs, &gl_attr)
+                .map(|(window, context)| GlWindow {
+                    window,
+                    context: Context { context },
+                })
+        }
+    }
+
+    /// Builds the given window along with the associated GL context, returning the pair as a
+    /// `GlWindow`.
+    ///
+    /// The context made can be shared with:
+    ///  - headless contexts made with the `shareable_with_windowed_contexts`
+    ///  flag set to `true`; and
+    ///  - contexts made when creating a `GlWindow`.
+    ///
+    /// You are not guaranteed to receive an error if you share a context with an
+    /// other context which you're not permitted to share it with, as according
+    /// to:
+    ///  - the restrictions stated by us above; and
+    ///  - the restrictions imposed on you by the platform your application runs
+    ///  on. (Please refer to `README-SHARING.md`)
+    ///
+    /// Failing to follow all the context sharing restrictions imposed on you
+    /// may result in unsafe behavior.
+    ///
+    /// Error should be very rare and only occur in case of permission denied,
+    /// incompatible system out of memory, etc.
+    pub unsafe fn new_shared(
         window_builder: WindowBuilder,
         context_builder: ContextBuilder,
         events_loop: &EventsLoop,
@@ -336,8 +406,8 @@ impl GlWindow {
         let gl_attr = gl_attr.map_sharing(|ctxt| &ctxt.context);
         platform::Context::new(window_builder, events_loop, &pf_reqs, &gl_attr)
             .map(|(window, context)| GlWindow {
-                window: window,
-                context: Context { context: context },
+                window,
+                context: Context { context },
             })
     }
 
@@ -349,6 +419,36 @@ impl GlWindow {
     /// Borrow the inner GL `Context`.
     pub fn context(&self) -> &Context {
         &self.context
+    }
+
+    /// Swaps the buffers in case of double or triple buffering.
+    ///
+    /// You should call this function every time you have finished rendering, or the image may not
+    /// be displayed on the screen.
+    ///
+    /// **Warning**: if you enabled vsync, this function will block until the next time the screen
+    /// is refreshed. However drivers can choose to override your vsync settings, which means that
+    /// you can't know in advance whether `swap_buffers` will block or not.
+    pub fn swap_buffers(&self) -> Result<(), ContextError> {
+        self.context.context.swap_buffers()
+    }
+
+    /// Returns the pixel format of the main framebuffer of the context.
+    pub fn get_pixel_format(&self) -> PixelFormat {
+        self.context.context.get_pixel_format()
+    }
+
+    /// Resize the context.
+    ///
+    /// Some platforms (macOS, Wayland) require being manually updated when their window or
+    /// surface is resized.
+    ///
+    /// The easiest way of doing this is to take every `Resized` window event that
+    /// is received with a `LogicalSize` and convert it to a `PhysicalSize` and
+    /// pass it into this function.
+    pub fn resize(&self, size: dpi::PhysicalSize) {
+        let (width, height) = size.into();
+        self.context.context.resize(width, height);
     }
 }
 
@@ -365,20 +465,95 @@ impl GlContext for Context {
         self.context.get_proc_address(addr)
     }
 
-    fn swap_buffers(&self) -> Result<(), ContextError> {
-        self.context.swap_buffers()
-    }
-
     fn get_api(&self) -> Api {
         self.context.get_api()
     }
+}
 
-    fn get_pixel_format(&self) -> PixelFormat {
-        self.context.get_pixel_format()
+impl Context {
+    /// Builds the given GL context
+    ///
+    /// Contexts made with the `shareable_with_windowed_contexts` flag set to
+    /// `true` can be shared with:
+    ///  - contexts made with that flag set to `true`; and
+    ///  - contexts made when creating a `GlWindow`.
+    ///
+    /// If the flag is set to `false` on the other hand, the context should only
+    /// be shared with other contexts made with the flag set to `false`.
+    ///
+    /// Some platforms might not implement contexts which aren't shareable with
+    /// windowed contexts. If so, those platforms will fallback to making a
+    /// contexts which are shareable with windowed contexts.
+    ///
+    /// You are not guaranteed to receive an error if you share a context with an
+    /// other context which you're not permitted to share it with, as according
+    /// to:
+    ///  - the restrictions stated by us above; and
+    ///  - the restrictions imposed on you by the platform your application runs
+    ///  on. (Please refer to `README-SHARING.md`)
+    ///
+    /// Failing to follow all the context sharing restrictions imposed on you
+    /// may result in unsafe behavior.
+    ///
+    /// This safe variant of `new_shared` will panic if you try to share it with
+    /// an existing context.
+    ///
+    /// Error should be very rare and only occur in case of permission denied,
+    /// incompatible system, out of memory, etc.
+    pub fn new(
+        el: &winit::EventsLoop,
+        context_builder: ContextBuilder,
+        shareable_with_windowed_contexts: bool,
+    ) -> Result<Self, CreationError>
+    {
+        let ContextBuilder { pf_reqs, gl_attr } = context_builder;
+        let gl_attr = gl_attr.map_sharing(|_ctxt| panic!("Context sharing is not allowed when using `new()`. Please instead use `new_shared()`."));
+        
+        // Not all platforms support context sharing yet, when they do, their 
+        // `new.*` functions should be marked unsafe.
+        #[allow(unused_unsafe)] 
+        unsafe {
+            platform::Context::new_context(el, &pf_reqs, &gl_attr, shareable_with_windowed_contexts)
+                .map(|context| Context { context })
+        }
     }
 
-    fn resize(&self, width: u32, height: u32) {
-        self.context.resize(width, height);
+    /// Builds the given GL context
+    ///
+    /// Contexts made with the `shareable_with_windowed_contexts` flag set to
+    /// `true` can be shared with:
+    ///  - contexts made with that flag set to `true`; and
+    ///  - contexts made when creating a `GlWindow`.
+    ///
+    /// If the flag is set to `false` on the other hand, the context should only
+    /// be shared with other contexts made with the flag set to `false`.
+    ///
+    /// Some platforms might not implement contexts which aren't shareable with
+    /// windowed contexts. If so, those platforms will fallback to making a
+    /// contexts which are shareable with windowed contexts.
+    ///
+    /// You are not guaranteed to receive an error if you share a context with an
+    /// other context which you're not permitted to share it with, as according
+    /// to:
+    ///  - the restrictions stated by us above; and
+    ///  - the restrictions imposed on you by the platform your application runs
+    ///  on. (Please refer to `README-SHARING.md`)
+    ///
+    /// Failing to follow all the context sharing restrictions imposed on you
+    /// may result in unsafe behavior.
+    ///
+    /// Error should be very rare and only occur in case of permission denied,
+    /// incompatible system, out of memory, etc.
+    pub unsafe fn new_shared(
+        el: &winit::EventsLoop,
+        context_builder: ContextBuilder,
+        shareable_with_windowed_contexts: bool,
+    ) -> Result<Self, CreationError>
+    {
+        let ContextBuilder { pf_reqs, gl_attr } = context_builder;
+        let gl_attr = gl_attr.map_sharing(|ctxt| &ctxt.context);
+        platform::Context::new_context(el, &pf_reqs, &gl_attr, shareable_with_windowed_contexts)
+            .map(|context| Context { context })
     }
 }
 
@@ -395,20 +570,8 @@ impl GlContext for GlWindow {
         self.context.get_proc_address(addr)
     }
 
-    fn swap_buffers(&self) -> Result<(), ContextError> {
-        self.context.swap_buffers()
-    }
-
     fn get_api(&self) -> Api {
         self.context.get_api()
-    }
-
-    fn get_pixel_format(&self) -> PixelFormat {
-        self.context.get_pixel_format()
-    }
-
-    fn resize(&self, width: u32, height: u32) {
-        self.context.resize(width, height);
     }
 }
 
@@ -431,6 +594,8 @@ pub enum CreationError {
     NoAvailablePixelFormat,
     PlatformSpecific(String),
     Window(WindowCreationError),
+    /// We received two errors, instead of one.
+    CreationErrorPair(Box<CreationError>, Box<CreationError>),
 }
 
 impl CreationError {
@@ -438,16 +603,16 @@ impl CreationError {
         match *self {
             CreationError::OsError(ref text) => &text,
             CreationError::NotSupported(text) => &text,
-            CreationError::NotSupported(_) => "Some of the requested attributes are not supported",
             CreationError::NoBackendAvailable(_) => "No backend is available",
             CreationError::RobustnessNotSupported => "You requested robustness, but it is \
                                                       not supported.",
             CreationError::OpenGlVersionNotSupported => "The requested OpenGL version is not \
                                                          supported.",
             CreationError::NoAvailablePixelFormat => "Couldn't find any pixel format that matches \
-                                                      the criterias.",
+                                                      the criteria.",
             CreationError::PlatformSpecific(ref text) => &text,
             CreationError::Window(ref err) => std::error::Error::description(err),
+            CreationError::CreationErrorPair(ref _err1, ref _err2) => "Received two errors."
         }
     }
 }
@@ -455,6 +620,16 @@ impl CreationError {
 impl std::fmt::Display for CreationError {
     fn fmt(&self, formatter: &mut std::fmt::Formatter) -> Result<(), std::fmt::Error> {
         formatter.write_str(self.to_string())?;
+
+        if let CreationError::CreationErrorPair(ref e1, ref e2) = *self {
+            write!(formatter, " Error 1: \"")?;
+            e1.fmt(formatter)?;
+            write!(formatter, "\"")?;
+            write!(formatter, " Error 2: \"")?;
+            e2.fmt(formatter)?;
+            write!(formatter, "\"")?;
+        }
+
         if let &CreationError::NotSupported(msg) = self {
             write!(formatter, ": {}", msg)?;
         }
@@ -488,6 +663,8 @@ impl From<WindowCreationError> for CreationError {
 /// Error that can happen when manipulating an OpenGL context.
 #[derive(Debug)]
 pub enum ContextError {
+    /// General platform error.
+    OsError(String),
     IoError(io::Error),
     ContextLost,
 }
@@ -496,6 +673,7 @@ impl ContextError {
     fn to_string(&self) -> &str {
         use std::error::Error;
         match *self {
+            ContextError::OsError(ref string) => string,
             ContextError::IoError(ref err) => err.description(),
             ContextError::ContextLost => "Context lost"
         }
@@ -585,7 +763,7 @@ pub enum Robustness {
     /// The driver doesn't check anything. This option is very dangerous. Please know what you're
     /// doing before using it. See the `GL_KHR_no_error` extension.
     ///
-    /// Since this option is purely an optimisation, no error will be returned if the backend
+    /// Since this option is purely an optimization, no error will be returned if the backend
     /// doesn't support it. Instead it will automatically fall back to `NotRobust`.
     NoError,